cli: make payments utilities local to route module
