+++ raiden-cli/src/types.ts
@@ -18,20 +18,3 @@ export interface Cli {
   app?: Express;
   server?: Server;
 }
-
-export enum ApiPaymentEvents {
-  sent = 'EventPaymentSentSuccess',
-  received = 'EventPaymentReceivedSuccess',
-}
-
-export interface ApiPayment {
-  event: ApiPaymentEvents;
-  initiator_address: string;
-  target_address: string;
-  token_address: string;
-  amount: string;
-  identifier: string;
-  secret: string;
-  secret_hash: string;
-  log_time: string;
-}
